+++ Examples/ObjC/Sources/ContentCards-Custom-UI/CardsInfoViewController.h
@@ -1,9 +1,10 @@
 @import UIKit;
+@import BrazeKit;
 
 @class BRZContentCardRaw;
 
 @interface CardsInfoViewController : UITableViewController
 
-- (instancetype)initWithCards:(NSArray<BRZContentCardRaw *> *)cards;
+- (instancetype)initWithBraze:(Braze *)braze;
 
 @end